Add configurable minimum permissions for triggers
For private repositories people with `Read` role (aka `pull` permission) are seen as collaborators
by the GitHub API. This means, they were able trigger builds with older versions of the plugin.

This change fixes this, and also adds the option to limit triggering of builds to repository admins.

The existing `allowUntrusted` toggle is dropped from the UI.
This change adds a new drop-down `Minimum Permissions on repository to trigger the build` with
these options:
- Only users with admin permission
- Only users that can push to the repository (default)
- Allow untrusted users to trigger the build